Create playRound function
I first converted the humanChoice parameter to lowercase, Then I used a switch statement for the rock, paper, and scissors values. Each case has if else statements that checks if human beat the computer. The winners are displayed and the scores are updated in each if statement except for else which is a draw